TRY-33 yaml config einlesen (#8)
* add config.yaml powered by standardisierungsausschuss

* TRY-33 add serde and serde_yaml libs to parse YAML files

* TRY-33 modified settings.rs to match the config.yaml structure

* TRY-33 config.yaml can be read now

* TRY-33 update

* TRY-33 update object.rs to load obj files

* TRY-33 delete unused imports

* TRY-37 new model

* TRY-33 passt

* TRY-33 commented a bunch of shit

---------

+++ src/scene/settings.rs
@@ -1,5 +1,13 @@
-#[derive(Debug, Clone, PartialEq)]
+use serde::{Deserialize, Serialize};
+
+// This struct represents settings for a program and is both serializable and deserializable
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Settings {
-    max_bounces: u32,
-    samples: u32,
+    
+    // This field specifies the maximum number of bounces for ray tracing or similar operations
+    max_bounces: usize,
+
+    // This field represents the number of samples used in rendering or some other operation
+    samples: usize,
 }
+
